[ts-sdk] Introduce SDK Connection class (#8088)
This is an idea to create a dedicated class to represent a `connection`,
meaning a bag of configuration related to how to connect to a network. I
thought about naming this `Network` but I thought that could be a little
confusing. Maybe something like `Endpoints` would be a better class
name? Open to suggestions on the naming.