[narwhal] additional consensus tests (#8479)
Small refactoring in error handling and certificate processing

Introduced more fine grained error handling when inserting a certificate in the DAG for equivocation protection. Also, refactored the process_certificate method to return not just the committed subdags, but an enum with the outcome of the processing to distinguish the various cases. Also two new tests have been introduced for Bullshark.