Remove DevX roadmap, update Sui Readme (#8701)
## Description 

Docs only update that includes:
 * Removed the DevX roadmap as we were not updating it regularly.
* Updated the Sui Readme to remove duplicative links, copyedit, and
added links to Sui Foundation and Sui Resources.

## Test Plan 

Manually on the docs staging server.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

This is a doc-only that removes the DevX roadmap and includes updates
for style consistency.