[TransactionManager] reduce time spent while holding lock in `enqueue()` (#8517)
Although `TransactionManager` may not be the immediate bottleneck during
Sui node catchup / checkpoint execution, it seems worthwhile to make
sure the contention is as low as possible.

This PR updates `TransactionManager::enqueue()` to only hold the
internal lock while updating internal in-memory state. The high level
logic is:
1. Without lock: Get versions for all input objects, reading them from
shared locks table if necessary. Then find input objects that are still
unavailable.
2. With lock: update internal state for pending certificates and missing
input objects.
3. Without lock: Re-check missing input objects to see if any of them
exists now.
4. With lock: call `objects_available()` with newly available objects.

In the ideal and likely common case, steps 3 and 4 are skipped because
no missing input object is found at step 1. The worst case scenario is
that an object's existence is checked twice: at steps 1 and 3.

Also, `TransactionManager` no longer takes tx lock in `enqueue()`,
because it seems unnecessary:
- Transactions may have multiple executions after crash recovery, when
they are recovered from both the recovery log and `pending_execution`
table. But this seems benign.
- Missing input objects table does not need this lock to ensure
available objects do not remain in the table forever.