Gas budget in SUI (#8676)
## Description 

Change Gas Budget to be in sui rather then gas units, that removes the
concept of gas units from any user related API.
Internally we move to gas units as soon as we make `SuiGasStatus` so the
change is very minimal.

The format of the `TransactionData` is not changed (`u64`), so
effectively this is not a breaking change as a format but it requires
clients to think about how they use gas budget.

Most of our codebase (tests) use a gas price of 1 (DUMMY_GAS_PRICE), so
GAS UNITS or MIST are the same, which explains why almost nothing broke.
Locally I changed the DUMMY_GAS_PRICE to be 2 and some tests broke,
however I verified that they were all using constants that had to be
changed. So this seems to be really the only change needed.
Tests pass so we are good.

I need to produce documentation about this change - and others I have -
but given we are not going out with this in devnet, that can wait a day
or two (doing it as soon as I move all blockers from my plate).

@Jordan-Mysten @gegaowp @patrickkuo @666lcz please let me know if you
see issues or we can move this forward

## Test Plan 

Existing test 

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes