[ledger integration] add ledgerjs-hw-app-sui library as an SDK (#8788)
## Description 

This PR adds the `ledgerjs-hw-app-sui` library from Obsidian as an SDK
under our repository since we'll be maintaining and owning it going
forward. I changed around and added some project configurations (e.g.
ESLint, Prettier, Vitest, etc.), but all of the code is pretty much
strictly copy and pasted over. I confirmed with the Obsidian team that
it's okay to move this code over and apply our software licenses.

There are some improvements and clean-up to be done in the SDK itself,
but I figure this is a good starting point since it's at least in a
working state. As follow-ups, we'll do some KTLO work to get the library
in better shape and add more comprehensive test coverage (i.e., unit
tests and E2E tests that interact with
[Speculos](https://github.com/LedgerHQ/speculos)).


## Test Plan 
- All `npm` scripts run successfully

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
N/A