Fix CI failure by upgrading remove_dir_all to 0.8.0 (#8859)
## Description 

`cargo deny` in CI is
[failing](https://github.com/MystenLabs/sui/actions/runs/4333210140/jobs/7566330958)
due to: https://github.com/advisories/GHSA-mc8h-8q98-g5hr. This fix is
to upgrade the `temp` package to a new version that does not have this
vulnerability.

## Test Plan 

CI

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes