[Narwhal] Remove Header::signature (#8932)
## Description 

Besides verifying the header is indeed published by the author, there
does not seem to other use cases for Header::signature. But the anemo
`request_vote()` handler already validates that the header author
matches the peer that sends out the request. So the `signature` field
seems redundant after anemo integration.