[types] reduce object size by special-casing framework types
Every Move object carries a `StructTag` and these are not small, especially for generic types. For example, a SUI coin's `StructTag` is `0x2::coin::Coin<0x2::sui::SUI>`, which is 84 bytes. The total size of a coin object is 214 bytes, so the type is ~40% of the total object size! This is especially problematic given that we charge by the byte for object storage, and a simple payment operation will require creating a new coin object. We want to make the storage footprint of coins and payments as low as possible.

This PR introduces `MoveType`, which is a wrapper around `StructTag` that special-cases a few common Sui Framework types (just gas tokens, `StakedSui`, and `Coin<T>`'s) for now. This lowers the storage footprint of gas tokens by 84 bytes, `StakedSui` by 53 bytes, and `Coin<T>` by 41 bytes. We can add more types (`Collectible<T>`? user-defined types) in the future if they become popular enough to justify the space savings.

The runtime representation of objects inside the Sui adapter also uses this new type representation to reduce the need to convert to/from `StructTag` and lower the memory footprint a bit, but a lot more could surely be done here.