[wallet-ext] spruce up the full screen wallet view and fix bug where overlays were being positioned incorrectly (#9032)
## Description 

This change does the following:
- Updates the styling of the wallet in full-screen mode (now it's
centered, rounded, and has a background image)
 - Removes the `limitToPopUpSize` prop size from the `Layout` component
- Fixes an issue where overlays were being incorrectly positioned in
full-screen mode because the overlay was using `fixed` positioning. I
think the more proper solution here is to portal the Overlay component
and use absolute positioning instead of trying to manually position the
Overlay on top of the wallet

Figma link:
https://www.figma.com/file/xu2VnyPAuruNi74UQzGigW/02-Wallet?node-id=3299-58195&t=Vwa9kJdV50yJjlma-0

Full-screen view:
https://capture.dropbox.com/xHqgyvTfh3WNtBqa

Pop-up view:
https://capture.dropbox.com/L79ke8MwGEeQdyKy

## Test Plan 
- Manual testing
  - Covered mostly every flow in the wallet in the browser window
  - Did the same thing as above for the wallet as a pop-up

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
N/A