client publish: gather dependencies (#9086)
## Description 

Changes the internal logic in `client publish` to actually collect
package dependencies in a partitioned set of `published`/`unpublished`/`invalid`
packages based on `published-at` values. We now separate error promotion for `unpublished` packages
when `--with-unpublished-dependencies` is not set (no user-visible
change) and `invalid` packages are hard errors.

## Test Plan 

Updated existing tests.