[wallet-ext] import and save Ledger accounts to the wallet (#9165)
## Description 

This PR splits the `Account` class into separate classes for each
account type: `LedgerAccount`, `ImportedAccount`, and `DerivedAccount`.
I think there are two options to go about supporting signing
transactions with Ledger in the wallet:
1. Make a keypair with async resolution that can connect to the user's
Ledger device from the background script
2. Remove the concept of a keypair for Ledger accounts and sign
transactions using a Ledger signer in the content script

While I think option 1 is possibly the more "proper" solution based on
some brief discussions with @Jordan-Mysten, there's added complexity to
updating all usages of the `Keypair` interface in the SDK (or creating
some discriminated union where `type Keypair =
KeypairWithAsyncResolution | KeypairWIthSyncResolution`). There's also
the matter of initializing a `SuiLedgerClient` instance from the
background script which has added complexity (how do we handle errors
when the user disconnects their device and then tries to sign a
transaction?).

So this PR implements the initial work for option 2 which is to remove
the concept of keypairs for Ledger accounts. While this might not be the
most ideal, I think it makes sense, and having distinct Account classes
allows us to be very explicit in code with what is or isn't possible for
an account (e.g., can I export the private key to this account?). If
option 1 might be more feasible or if there are concerns with this
approach, I'm very glad to hear any feedback 😄.

Part 1 (this diff): https://github.com/MystenLabs/sui/pull/9165
Part 2 (add the ability to save/load Ledger accounts):
https://github.com/MystenLabs/sui/pull/9166
Part 3 (hook-up UI to save imported Ledger accounts):
https://github.com/MystenLabs/sui/pull/9167
Part 4 (add account badge for Ledger accounts):
https://github.com/MystenLabs/sui/pull/9168

## Test Plan 
- Manual testing (No visible user changes and everything works the same
as before)

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
N/A