Version CheckpointContents (#9040)
## Description 

Stacked on #8972 

Version CheckpointContents.

Note that I didn't use the usual `#[enum_dispatch]` approach in this
case because I think its unlikely we are going to extend this format any
time soon. We can always add that in later if we need it.

## Test Plan 

```cargo simtest```

---
If your changes are not user-facing and not a breaking change, you can skip the following section. Otherwise, please indicate what changed, and then add to the Release Notes section as highlighted during the release process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes