move epoch_duration_ms from NodeConfig to on-chain (#9195)
## Description 

This PR moves the length of each epoch `epoch_duration_ms` to live
on-chain, instead of in the node config. At the beginning of each epoch,
we read the system state from on-chain and use the value we read to
populate `EpochStartSystemState`.

## Test Plan 

Making sure the existing tests pass.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [x] breaking change for a client SDKs
- [x] breaking change for FNs (FN binary must upgrade)
- [x] breaking change for validators or node operators (must upgrade
binaries)
- [x] breaking change for on-chain data layout
- [x] necessitate either a data wipe or data migration

### Release notes