[sui-cli][test fix] Fix flakiness around ordering of objects (#9233)
This solves another issue in the same test where objects were being
selected from a list of objects that were all expected to be coins. This
updates the test to first filter out all of the coins and only select
from these.

## Test Plan

Ran the test 20 times in a row after the fix with no failures.