[ts-sdk] Automatic gas budgeting (#9274)
## Description 

This changes the default budgeting for transactions in the SDK, so that
it's automatically derived. We currently do this with the following
simple approach:
1. Dry run the transaction with no gas payment and maximum gas budget.
2. Take the gas budget and add some overhead budget based on the amount
of coins being provided as payment.

I noticed that dry run is not super useful for us because of this, but
I'll work with @dariorussi on reworking dry-run so that this is a little
easier to reason about.

If consumers want to display the gas budget, they can build the
transaction to bytes, then use the `transactionData` to get the computed
budget out.

## Test Plan 

Tests should still pass with dynamic budgeting.