[wallet-ext] add signing capabilities for Ledger accounts (#9268)
## Description 

This PR adds signing capabilities to the wallet extension for Ledger
accounts. Will update this with a video once I get a Ledger application
build that has blind signing enabled (and of course, I won't land until
then either).

I'll say that I don't *love* this approach, but I tried several other
approaches that just didn't work out with the current architecture of
the wallet. I left some comments in the code explaining these issues,
but basically, the Ledger signer needs to live in the content script of
the wallet extension and be initialized on demand. If someone has other
ideas, I'd be glad to hear them but this PR also needs to land today if
we want to make our Ledger deadline.

## Test Plan 
- Manual testing (this is a lie atm but I'll be sure to test all of this
when possible)
- Signed transactions for sending money from one Ledger account to
another
- Successfully responded to a request to sign a transaction with my
Ledger account
  - Tested signing with multiple Ledger accounts
- Tested out some error scenarios (Ledger device not connected when
signing, signing with a connection open)

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
N/A